MDX用のコンポーネントを分割
